Fix/safe indexing bound check (#37)
* [ias-verify] fix potential index out of bound error in `safe_indexing`

* [ias-verify] add start equals end test

* [ias-verify] rename test
